+++ vargrid.go
@@ -941,7 +941,7 @@ func (c *Cell) loadData(data *CTMData, k int) error {
 		}
 	}
 	if sum := floats.Sum(fractions); sum < 0.9 {
-		return fmt.Errorf("there is not CTM data overlapping at least 90%% of the InMAP cell at %+v", c.Centroid())
+		return fmt.Errorf("there is not CTM data overlapping at least 90 percent of the InMAP cell at %+v", c.Centroid())
 	}
 	for i, ctmcell := range ctmcells {
 		ctmrow := ctmcell.Row